+++ src/PhpSpreadsheet/Style/NumberFormat/DateFormatter.php
@@ -3,6 +3,8 @@
 namespace PhpOffice\PhpSpreadsheet\Style\NumberFormat;
 
 use PhpOffice\PhpSpreadsheet\Shared\Date;
+use PhpOffice\PhpSpreadsheet\Shared\StringHelper;
+use Throwable;
 
 class DateFormatter
 {
@@ -161,7 +163,11 @@ class DateFormatter
         $callback = [self::class, 'escapeQuotesCallback'];
         $format = (string) preg_replace_callback('/"(.*)"/U', $callback, $format);
 
-        $dateObj = Date::excelToDateTimeObject($value);
+        try {
+            $dateObj = Date::excelToDateTimeObject($value);
+        } catch (Throwable) {
+            return StringHelper::convertToString($value);
+        }
         // If the colon preceding minute had been quoted, as happens in
         // Excel 2003 XML formats, m will not have been changed to i above.
         // Change it now.